ch4/ofi: refactor huge functions to separate file
It is not critical to inline functions related to huge messages as they
are bandwidth dominated. Move them into ofi_huge.c for better context.
